RiscVOS: Added new build script and cleaned the repo.
The build script can be found at scripts/build.py
This script makes it easier to compile and run the kernel, as it
downloads the riscv toolchain, clang and other needed tools and
libraries.

It has a lot of scripts to make ones life easier. Just run
`scripts/build.py -l` to list all the targets and try it out.