RiscVOS: Traps are now handled by the trap_vector
A new special virtual memory address was added to each process.
Every process now has its own TrapFrame* as well as KernelTrapFrame*
which are both accessible only in the S-mode. The KernelTrapFrame* will
be changed every time the process runs on different hart.

For this to work, I want to implement attaching and detaching from the
hart. Each hart now has its unique id, the cpu_id, in the
KernelTrapFrame*.