RiscVOS: ErrorOr & Vector should be working now!
I've edited the ErrorOr implementation by adding the Variant class. I've
also added a new type system which is the replacement for std typeid().
It can be used as so TypeID<void>::get().

I've also implemented String::split(String) methods, that takes a String
separator and uses that to split the string into an ArrayList<String> of
substrings.