RiscVOS/VirtIO: VirtIO devices are now being initialized
Each VirtIO device is first checked if connected. If connected, the
device will be initialized using the Device Initialization procedure.

Another change that is not related to VirtIO is contained in this commit
too. If there is an interrupt, the ISR will finally return in U-mode to
the process which was running during the interrupt. Of course, if it
was Timer interrupt, next process will be scheduled.