RiscVOS/FS + Init: I've added FATFS::open and fixed sp bug
There was 1 page allocated for the kernel stack, but we set it to the
top of the stack, so when there was a new stack allocation, it has been
slowly corrupting the data around it.

I've set the kernel sp back to the initial sp from linker script and for
other kernel thread's, I've changed it so there are 4 pages allocated
for these and they will start at the bottom of the stack and can now
freely grow.

I thought that when the stack grows here, the address increases, but
it's actually the other way around - the address decreases. :^)